Phase-2/backend/requirements.txt
fastapi
uvicorn
sqlmodel
psycopg2-binary
python-jose[cryptography]
pydantic-settings
python-dotenv
